+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 Notable changes.
 
+## March 2026
+
+### [0.84.1]
+- Bump tar from 7.5.10 to 7.5.11 due to [CVE-2026-31802](https://github.com/advisories/GHSA-9ppj-qmqm-q256). (https://github.com/devcontainers/cli/pull/1174)
+
+### [0.84.0]
+- Dependencies update. (https://github.com/devcontainers/cli/pull/1167)
+
+## February 2026
+
+### [0.83.3]
+- Bump tar from 7.5.7 to 7.5.8. (https://github.com/devcontainers/cli/pull/1160)
+
+### [0.83.2]
+- Improved logging for image inspect errors. (https://github.com/devcontainers/cli/pull/1152)
+
+### [0.83.1]
+- Bump tar from 7.5.6 to 7.5.7. (https://github.com/devcontainers/cli/pull/1140)
+
+### [0.83.0]
+- Add install script. (https://github.com/devcontainers/cli/pull/1142)
+- Remove request body limit. (https://github.com/devcontainers/cli/pull/1141)
+- Add BUILDKIT_INLINE_CACHE for container Feature path. (https://github.com/devcontainers/cli/pull/1135)
+
 ## January 2026
 
 ### [0.82.0]